19.06
In test2 fixed the logic behind sharpe and sortino ratio calculation, run experiments with models and expanded the LSTM model with a more complex function. ReLU works the best in comparison to Gelu, tanh and leaky relu activations.

Started working on new models in test3, currently the predictions are constant for each model type what indicates issues with the new training routine.